Split scheme.rs.
